+++ tests/update/query.test.js
@@ -140,7 +140,7 @@ describe("query function", () => {
     ];
     const consumer = { id: 1, nickname: 'consumer' };
     const { query, args } = entity.query.update(chunk, consumer);
-    expect(query).toBe(`UPDATE public.persons SET name = CASE WHEN id = $1 THEN $3 WHEN id = $2 THEN $4 ELSE name END, age = CASE WHEN id = $1 THEN $5 WHEN id = $2 THEN $6 ELSE age END, "consumerId" = CASE WHEN id = $1 THEN $7 WHEN id = $2 THEN $8 ELSE "consumerId" END, "consumerName" = CASE WHEN id = $1 THEN $9 WHEN id = $2 THEN $10 ELSE "consumerName" END WHERE id IN ($1, $2)`);
+    expect(query).toBe(`UPDATE public.persons SET name = CASE WHEN id = $1 THEN $3 WHEN id = $2 THEN $4 ELSE name END, age = CASE WHEN id = $1 THEN $5 WHEN id = $2 THEN $6 ELSE age END, updaterid = CASE WHEN id = $1 THEN $7 WHEN id = $2 THEN $8 ELSE updaterid END, name = CASE WHEN id = $1 THEN $9 WHEN id = $2 THEN $10 ELSE name END WHERE id IN ($1, $2)`);
     expect(args).toEqual([
       1, 2, 'John', 'Henry', 30, 40, 1, 1, 'consumer', 'consumer'
     ]);